Flatten InMemoryDownload and LibraryDownload into their parents.
